@@ -1,17 +1,18 @@
 {
   "name": "rehype-annotate",
-  "version": "0.1.0",
-  "description": "rehype plugin to match W3C Web Annotations to source locations",
+  "version": "1.0.0",
+  "description": "ReHype plugin to match W3C Web Annotations to source locations",
   "license": "Apache-2.0",
+  "type": "module",
   "scripts": {
-    "lint": "tsc --checkJs --allowJs --noEmit --lib es2015 --moduleResolution node  index.js && healthier && prettier --write --plugin-search-dir=. '**/*.{js,json,css,yml,svelte}'",
+    "lint": "tsc --module es2020 --checkJs --allowJs --noEmit --lib es2015 --moduleResolution node  index.js && healthier && prettier --write --plugin-search-dir=. '**/*.{js,json,css,yml,svelte}'",
     "pretest": "npm run lint",
     "prepublishOnly": "npm run lint",
-    "test": "tap tests/index.js --coverage-report=text --coverage-report=lcov ",
+    "test": "c8 tap --no-cov tests/index.js && c8 report --reporter=lcov",
     "snapshots": "TAP_SNAPSHOT=1 npm run test"
   },
   "engines": {
-    "node": ">10"
+    "node": ">14"
   },
   "healthier": {
     "ignore": [
@@ -20,38 +21,39 @@
     ]
   },
   "dependencies": {
-    "@types/node": "^12.12.11",
-    "aria-attributes": "^1.3.1",
+    "@types/node": "^16.0.1",
+    "aria-attributes": "^2.0.0",
     "deepmerge": "^4.2.2",
-    "diff-match-patch": "1.0.4",
-    "glob": "7.1.6",
-    "hast-util-sanitize": "^2.0.1",
-    "hast-util-select": "^3.0.1",
-    "hast-util-to-string": "^1.0.2",
-    "hastscript": "^5.1.1",
-    "html-element-attributes": "^2.2.0",
-    "property-information": "^5.3.0",
-    "rehype-parse": "^6.0.2",
-    "rehype-raw": "^4.0.1",
-    "rehype-sanitize": "^3.0.0",
-    "rehype-stringify": "^6.0.1",
-    "remark-parse": "^7.0.2",
-    "remark-rehype": "^5.0.0",
+    "diff-match-patch": "^1.0.5",
+    "glob": "^7.1.7",
+    "hast-util-sanitize": "^4.0.0",
+    "hast-util-select": "^5.0.0",
+    "hast-util-to-string": "^1.0.4",
+    "hastscript": "^7.0.1",
+    "html-element-attributes": "^3.0.0",
+    "property-information": "^6.0.1",
+    "rehype-parse": "^7.0.1",
+    "rehype-raw": "^5.1.0",
+    "rehype-sanitize": "^4.0.0",
+    "rehype-stringify": "^8.0.0",
+    "remark-parse": "^9.0.0",
+    "remark-rehype": "^8.1.0",
     "string-pixel-width": "^1.10.0",
-    "svg-element-attributes": "^1.3.0",
-    "to-vfile": "^6.0.0",
-    "unified": "^8.4.2",
-    "unist-util-is": "^4.0.1",
-    "unist-util-visit": "^2.0.1",
-    "unist-util-visit-parents": "^3.1.0",
+    "svg-element-attributes": "^2.0.0",
+    "to-vfile": "^7.1.0",
+    "unified": "^9.2.1",
+    "unist-util-is": "^5.1.0",
+    "unist-util-visit": "^3.1.0",
+    "unist-util-visit-parents": "^4.1.1",
     "universal-base64url": "^1.1.0",
-    "vfile-reporter": "^6.0.0"
+    "vfile-reporter": "^7.0.1"
   },
   "devDependencies": {
-    "healthier": "3.2.0",
-    "prettier": "1.19.1",
-    "tap": "^14.10.8",
-    "typescript": "^3.7.2"
+    "c8": "^7.7.3",
+    "healthier": "^4.0.0",
+    "prettier": "^2.3.2",
+    "tap": "^15.0.9",
+    "typescript": "^4.3.5"
   },
   "private": true
 }
