@@ -1,13 +1,11 @@
-const { select } = require("hast-util-select");
-const addPropsToNode = require("./add-props-to-node");
-
-module.exports = nodeSelector;
+import { select } from "hast-util-select";
+import { addPropsToNode } from "./add-props-to-node.js";
 
 /**
  *
  * @param {{tree: Object, value: string, annotation: Object, addProps: undefined | boolean}} param0 - selector options
  */
-function nodeSelector({ tree, value, annotation, addProps = true }) {
+export function nodeSelector({ tree, value, annotation, addProps = true }) {
   const node = select(value, tree);
   if (node && addProps) {
     addPropsToNode(node, annotation);
