@@ -1,8 +1,8 @@
 // const debug = require("../logger")("hast-annotations-match");
-const nodeSelector = require("./css-selector");
-const simpleXpathSelector = require("./simple-xpath-selector");
-const processPositions = require("./process-positions");
-const processQuotations = require("./process-quotations");
+import { nodeSelector } from "./css-selector.js";
+import { simpleXpathSelector } from "./simple-xpath-selector.js";
+import { processPositions } from "./process-positions.js";
+import { processQuotations } from "./process-quotations.js";
 const selectors = {
   XPathSelector: simpleXpathSelector,
   CssSelector: nodeSelector,
@@ -11,25 +11,23 @@ const selectors = {
       tree,
       value: "#" + value,
       annotation,
-      addProps
+      addProps,
     });
-  }
+  },
 };
-
-module.exports = getNode;
 
 /**
  *
  * @param {{tree: Object, selector: Object, annotation: Object}} param0 - selector options
  */
-function getNode({ tree, selector, annotation }) {
+export function getNode({ tree, selector, annotation }) {
   // Need to check `refinedBy`. If so and refining selector is quote or text-position then process using node as root tree
   if (selector.refinedBy) {
     const node = selectors[selector.type]({
       tree,
       value: selector.value,
       annotation,
-      addProps: false
+      addProps: false,
     });
     const target = { ...annotation.target, selector: selector.refinedBy };
     if (selector.refinedBy.type === "TextQuoteSelector") {
@@ -42,15 +40,15 @@ function getNode({ tree, selector, annotation }) {
         value: selector.refinedBy.value,
         annotation: {
           ...annotation,
-          target
-        }
+          target,
+        },
       });
     }
   } else {
     return selectors[selector.type]({
       tree,
       value: selector.value,
-      annotation
+      annotation,
     });
   }
 }
