@@ -1,14 +1,17 @@
-const addPropsToNode = require("./add-props-to-node");
+import { addPropsToNode } from "./add-props-to-node.js";
 
 // Based on simple-xpath-selector from https://github.com/tilgovi/simple-xpath-position/blob/master/src/xpath.js MIT license
-
-module.exports = simpleXpathSelector;
 
 /**
  *
  * @param {{tree: Object, value: string, annotation: Object, addProps: undefined | boolean}} param0 - selector options
  */
-function simpleXpathSelector({ tree, value, annotation, addProps = true }) {
+export function simpleXpathSelector({
+  tree,
+  value,
+  annotation,
+  addProps = true,
+}) {
   const node = fallbackResolve(value, tree);
   if (node && addProps) {
     addPropsToNode(node, annotation);
