@@ -1,5 +1,5 @@
 module.exports = function log(path) {
-  return function(...args) {
+  return function (...args) {
     if (process.env.NODE_ENV !== "production") {
       console.log(`${path}: `, ...args);
     }
