@@ -1,37 +1,41 @@
+import { toVFile } from "to-vfile";
+import { reporter } from "vfile-reporter";
+import { attacher } from "../index.js";
+import * as path from "path";
+import { createRequire } from "module";
+import { fileURLToPath } from "url";
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename).replace(process.cwd() + "/", "");
+const require = createRequire(import.meta.url);
 const tap = require("tap");
-const vfile = require("to-vfile");
 const unified = require("unified");
-const annotate = require("../");
 const parse = require("rehype-parse");
 const stringify = require("rehype-stringify");
-const report = require("vfile-reporter");
 const glob = require("glob");
-const path = require("path");
-
 // const fragment = { fragment: true };
 
-async function process(filepath, options) {
+async function processAnnotations(filepath, options) {
   let file;
-  const input = await vfile.read(filepath);
+  const input = await toVFile.read(filepath);
   try {
     file = await unified()
       .use(parse)
-      .use(annotate, options)
+      .use(attacher, options)
       .use(stringify)
       .process(input);
-    console.log(report(file));
+    console.log(reporter(file));
   } catch (err) {
     console.log(err);
-    console.error(report(err));
+    console.error(reporter(err));
   }
   return file;
 }
 
-tap.test("rehype-annotate", async t => {
+tap.test("rehype-annotate", async (t) => {
   const fixtures = glob.sync("fixtures/*.input.html", { cwd: __dirname });
   for (const fixture of fixtures) {
-    await t.test("basic " + fixture, async function(t) {
-      const result = await process(path.join(__dirname, fixture), {
+    await t.test("basic " + fixture, async function (t) {
+      const result = await processAnnotations(path.join(__dirname, fixture), {
         stimulus: fixture.includes("quote") || fixture.includes("xpath"),
         annotations: require("./" +
           fixture.replace("input.html", "annotations.json")).items,
@@ -39,15 +43,16 @@ tap.test("rehype-annotate", async t => {
         canonical: `https://canonical.example.com/tests/${fixture.replace(
           __dirname,
           ""
-        )}`
+        )}`,
       });
       const expected = String(
-        vfile.readSync(
+        toVFile.readSync(
           path.join(__dirname, fixture.replace("input.html", "output.html"))
         )
       );
       t.matchSnapshot(result.data.annotations, `${fixture} annotations`);
-      t.equal(String(result), expected);
+      // console.dir(result);
+      t.equal(String(result.contents), expected);
       t.end();
     });
   }
